Minimo Tema theme Venus
A minimial Drupal 8 theme

Now with Bootstrap 4.1

Jorge L Lopez
IT 390 - Spring 2018 Version - 1805.21
Copyright 2018

Theme is based on work for IT 390 DePaul and personal portfolio using. Theme is built for Drupal 8 Twig.

This theme is now in version 0.2.0, aka Venus, and utilizes a more advanced subset of features. Regions are now defined closely to actual layout