Minimo Tema theme< Jupiter
A minimial Drupal 8 TWIG theme
Now with Bootstrap 4.1

Jorge L Lopez
Copyright 2018

Theme started as a base theme for work in IT 390 DePaul. Theme and site will be the basis of personal portfolio. Theme is built for Drupal 8 Twig.

This theme is now in version 0.2.0, aka Venus, and utilizes a more advanced subset of features. Regions are now defined closely to actual layout

Merge remote-tracking branch 'origin/master'